add commands, usercode, and feature_bits to the nvcm. test does not work, fails at verifying nvcm array.

+++ Core/Src/crosslink.c
@@ -798,7 +798,7 @@ int fpga_configure(I2C_HandleTypeDef *hi2c, uint16_t DevAddress, GPIO_TypeDef *G
     return ret_status;
 }
 
-int fpga_nvcm_configure(I2C_HandleTypeDef *hi2c, uint16_t DevAddress, GPIO_TypeDef *GPIOx, uint16_t GPIO_Pin, const uint8_t *config_data, uint32_t data_size, uint32_t usercode, uint8_t feature_bits)
+int fpga_nvcm_configure(I2C_HandleTypeDef *hi2c, uint16_t DevAddress, GPIO_TypeDef *GPIOx, uint16_t GPIO_Pin, const uint8_t *config_data, uint32_t data_size, uint32_t usercode, uint16_t feature_bits)
 {
     int ret;
